Various changes because of test results - frequencies now double, was int - vex2ccf and ccf parser code now tuned
M    include/genFunctions.h
M    include/genPrms.h
M    include/DelayCorrection.h
M    TODO.txt
M    src/DelayCorrection.cc
M    src/sfxc_SC.cc
M    src/staPrms.cc
M    src/genFunctions.cc
M    src/genPrms.cc
M    utils/Makefile.am
M    utils/vex_parser/vex2ccf.cc
M    utils/vex_parser/y.tab.h


git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@172 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/DelayCorrection.h
@@ -82,9 +82,9 @@ class DelayCorrection
     double tbs;         //time between samples
     int    Nf;          //nr of frequencies in frequency scale
     double *fs;         //delta frequency in frequency scale, frequency scale
-    INT32  foffset;     //frequency offset in Hertz
-    INT32  bwfl;        //band width after filter 
-    INT32  startf;      //start frequency
+    double foffset;     //frequency offset in Hertz
+    double bwfl;        //band width after filter 
+    double startf;      //start frequency
     int    ovrfl;       //oversampling in filter
 
     int    n2fftcorr;   //FFT length in correlation
@@ -94,7 +94,7 @@ class DelayCorrection
     fftw_plan    planT2F;//plan for complex FFT Time to Frequeny
     fftw_plan    planF2T;//plan for complex FFT Frequency to Time
 
-    INT64 skyfreq;       //channel sky frequency
+    double skyfreq;       //channel sky frequency
 
     std::vector<DelayTable> delTbl;
